requirements.txt
cohere==4.26.1
langchain==0.0.294
pinecone_client==2.2.4
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.26.0
